refactor: break app into multiple files
