+++ run.py
@@ -1,0 +1,7 @@
+# run.py
+
+from wf_game_tracker.main_app import DeckTrackerApp
+
+if __name__ == "__main__":
+    app = DeckTrackerApp()
+    app.mainloop()
